dino discriminator

+++ trainer.py
@@ -23,7 +23,7 @@ def trainerThread(cfg, s2c = None, c2s = None):
     trainer = Trainer(
             accelerator='gpu', 
             devices=1,
-            precision=16,
+            precision= '16-mixed',
             # amp_backend ="apex", 
             # amp_level='O1',
             callbacks=[checkpoint_callback,lr_monitor],
